Coderefactoring (#342)
* use infoTextFontSize and adjust the class

* add license header

* adjust _createItemList()

* reorganize chart files into screen-specific folders
